+++ model/marshal_fastjson.go
@@ -135,6 +135,7 @@ func (v *Runtime) MarshalFastJSON(w *fastjson.Writer) error {
 }
 
 func (v *System) MarshalFastJSON(w *fastjson.Writer) error {
+	var firstErr error
 	w.RawByte('{')
 	first := true
 	if v.Architecture != "" {
@@ -146,6 +147,18 @@ func (v *System) MarshalFastJSON(w *fastjson.Writer) error {
 			w.RawString(prefix)
 		}
 		w.String(v.Architecture)
+	}
+	if v.Container != nil {
+		const prefix = ",\"container\":"
+		if first {
+			first = false
+			w.RawString(prefix[1:])
+		} else {
+			w.RawString(prefix)
+		}
+		if err := v.Container.MarshalFastJSON(w); err != nil && firstErr == nil {
+			firstErr = err
+		}
 	}
 	if v.Hostname != "" {
 		const prefix = ",\"hostname\":"
@@ -168,7 +181,7 @@ func (v *System) MarshalFastJSON(w *fastjson.Writer) error {
 		w.String(v.Platform)
 	}
 	w.RawByte('}')
-	return nil
+	return firstErr
 }
 
 func (v *Process) MarshalFastJSON(w *fastjson.Writer) error {
@@ -194,6 +207,14 @@ func (v *Process) MarshalFastJSON(w *fastjson.Writer) error {
 		w.RawString(",\"title\":")
 		w.String(v.Title)
 	}
+	w.RawByte('}')
+	return nil
+}
+
+func (v *Container) MarshalFastJSON(w *fastjson.Writer) error {
+	w.RawByte('{')
+	w.RawString("\"id\":")
+	w.String(v.ID)
 	w.RawByte('}')
 	return nil
 }
@@ -282,6 +303,10 @@ func (v *Span) MarshalFastJSON(w *fastjson.Writer) error {
 	}
 	w.RawString(",\"type\":")
 	w.String(v.Type)
+	if v.Action != "" {
+		w.RawString(",\"action\":")
+		w.String(v.Action)
+	}
 	if v.Context != nil {
 		w.RawString(",\"context\":")
 		if err := v.Context.MarshalFastJSON(w); err != nil && firstErr == nil {
@@ -306,6 +331,10 @@ func (v *Span) MarshalFastJSON(w *fastjson.Writer) error {
 			}
 		}
 		w.RawByte(']')
+	}
+	if v.Subtype != "" {
+		w.RawString(",\"subtype\":")
+		w.String(v.Subtype)
 	}
 	w.RawByte('}')
 	return firstErr