+++ module/apmgrpc/client_test.go
@@ -10,7 +10,7 @@ import (
 	"golang.org/x/net/context"
 	pb "google.golang.org/grpc/examples/helloworld/helloworld"
 
-	"go.elastic.co/apm"
+	"go.elastic.co/apm/apmtest"
 	"go.elastic.co/apm/transport/transporttest"
 )
 
@@ -18,7 +18,9 @@ func TestClientSpan(t *testing.T) {
 	tracer, transport := transporttest.NewRecorderTracer()
 	defer tracer.Close()
 
-	s, _, addr := newServer(t, nil) // no server tracing
+	serverTracer, serverTransport := transporttest.NewRecorderTracer()
+	defer serverTracer.Close()
+	s, _, addr := newServer(t, serverTracer)
 	defer s.GracefulStop()
 
 	conn, client := newClient(t, addr)
@@ -31,15 +33,23 @@ func TestClientSpan(t *testing.T) {
 	tracer.Flush(nil)
 	require.Zero(t, transport.Payloads())
 
-	tx := tracer.StartTransaction("name", "type")
-	ctx := apm.ContextWithTransaction(context.Background(), tx)
-	resp, err = client.SayHello(ctx, &pb.HelloRequest{Name: "birita"})
-	require.NoError(t, err)
-	assert.Equal(t, resp, &pb.HelloReply{Message: "hello, birita"})
-	tx.End()
+	_, clientSpans, _ := apmtest.WithTransaction(func(ctx context.Context) {
+		resp, err = client.SayHello(ctx, &pb.HelloRequest{Name: "birita"})
+		require.NoError(t, err)
+		assert.Equal(t, resp, &pb.HelloReply{Message: "hello, birita"})
+	})
 
-	tracer.Flush(nil)
-	spans := transport.Payloads().Spans
-	require.Len(t, spans, 1)
-	assert.Equal(t, "/helloworld.Greeter/SayHello", spans[0].Name)
+	require.Len(t, clientSpans, 1)
+	assert.Equal(t, "/helloworld.Greeter/SayHello", clientSpans[0].Name)
+	assert.Equal(t, "external", clientSpans[0].Type)
+	assert.Equal(t, "grpc", clientSpans[0].Subtype)
+
+	serverTracer.Flush(nil)
+	serverTransactions := serverTransport.Payloads().Transactions
+	require.Len(t, serverTransactions, 2)
+	for _, tx := range serverTransactions {
+		assert.Equal(t, "/helloworld.Greeter/SayHello", tx.Name)
+	}
+	assert.Equal(t, clientSpans[0].TraceID, serverTransactions[1].TraceID)
+	assert.Equal(t, clientSpans[0].ID, serverTransactions[1].ParentID)
 }